WEB-2073 extract textProps to separate file
