+++ src/card.tsx
@@ -3,7 +3,8 @@ import * as React from 'react';
 import Tag from './tag';
 import Stack from './stack';
 import Box from './box';
-import {Text2, Text, Text6, Text3, textProps} from './text';
+import {Text2, Text, Text6, Text3} from './text';
+import * as textProps from './text-props';
 import {Boxed, InternalBoxed} from './boxed';
 import ButtonGroup from './button-group';
 import Image from './image';
