+++ src/community/advanced-data-card.tsx
@@ -7,7 +7,8 @@ import Stack from '../stack';
 import * as styles from './advanced-data-card.css';
 import * as mediaStyles from '../image.css';
 import Divider from '../divider';
-import {Text2, Text, textProps} from '../text';
+import {Text2, Text} from '../text';
+import * as textProps from '../text-props';
 import {vars} from '../skins/skin-contract.css';
 import Box from '../box';
 import Touchable from '../touchable';
