+++ src/empty-state-card.tsx
@@ -4,7 +4,8 @@ import Box from './box';
 import {Boxed} from './boxed';
 import {useTheme} from './hooks';
 import Stack from './stack';
-import {Text2, Text, textProps} from './text';
+import {Text2, Text} from './text';
+import * as textProps from './text-props';
 import ButtonGroup from './button-group';
 import * as styles from './empty-state-card.css';
 import {vars} from './skins/skin-contract.css';
