+++ src/highlighted-card.tsx
@@ -3,7 +3,8 @@ import * as React from 'react';
 import {useIsInverseOrMediaVariant} from './theme-variant-context';
 import Stack from './stack';
 import {BaseTouchable} from './touchable';
-import {Text, Text2, textProps} from './text';
+import {Text, Text2} from './text';
+import * as textProps from './text-props';
 import {Boxed} from './boxed';
 import MaybeDismissable, {useIsDismissable} from './maybe-dismissable';
 import * as styles from './highlighted-card.css';
