+++ src/table.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import {getPrefixedDataAttributes} from './utils/dom';
 import * as styles from './table.css';
 import {vars} from './skins/skin-contract.css';
-import Text, {Text1, Text2, textProps} from './text';
+import Text, {Text1, Text2} from './text';
+import * as textProps from './text-props';
 import {InternalBoxed} from './boxed';
 import classNames from 'classnames';
 import Box from './box';
