+++ src/text-props.tsx
@@ -1,0 +1,72 @@
+// text props defined in a separate file to avoid exporting them from a client component.
+// If exported from a client component like "text.tsx" and in a RSC, the RSC will be forced to be rendered as a client component.
+
+export const text1 = {
+    mobileSize: 12,
+    mobileLineHeight: '16px',
+    desktopSize: 14,
+    desktopLineHeight: '20px',
+};
+
+export const text2 = {
+    mobileSize: 14,
+    mobileLineHeight: '20px',
+    desktopSize: 16,
+    desktopLineHeight: '24px',
+};
+
+export const text3 = {
+    mobileSize: 16,
+    mobileLineHeight: '24px',
+    desktopSize: 18,
+    desktopLineHeight: '24px',
+};
+
+export const text4 = {
+    mobileSize: 18,
+    mobileLineHeight: '24px',
+    desktopSize: 20,
+    desktopLineHeight: '28px',
+};
+
+export const text5 = {
+    mobileSize: 20,
+    mobileLineHeight: '24px',
+    desktopSize: 28,
+    desktopLineHeight: '32px',
+};
+
+export const text6 = {
+    mobileSize: 24,
+    mobileLineHeight: '32px',
+    desktopSize: 32,
+    desktopLineHeight: '40px',
+};
+
+export const text7 = {
+    mobileSize: 28,
+    mobileLineHeight: '32px',
+    desktopSize: 40,
+    desktopLineHeight: '48px',
+};
+
+export const text8 = {
+    mobileSize: 32,
+    mobileLineHeight: '40px',
+    desktopSize: 48,
+    desktopLineHeight: '56px',
+};
+
+export const text9 = {
+    mobileSize: 40,
+    mobileLineHeight: '48px',
+    desktopSize: 56,
+    desktopLineHeight: '64px',
+};
+
+export const text10 = {
+    mobileSize: 48,
+    mobileLineHeight: '56px',
+    desktopSize: 64,
+    desktopLineHeight: '72px',
+};
