+++ src/text.tsx
@@ -9,6 +9,7 @@ import {vars} from './skins/skin-contract.css';
 import * as styles from './text.css';
 import {VIVO_NEW_SKIN} from './skins/constants';
 import ScreenReaderOnly from './screen-reader-only';
+import * as textProps from './text-props';
 
 import type {ExclusifyUnion} from './utils/utility-types';
 import type {FontWeight} from './skins/types';
@@ -281,69 +282,6 @@ const getTextSizes = ({
             desktopLineHeight,
         };
     }
-};
-
-export const textProps = {
-    text1: {
-        mobileSize: 12,
-        mobileLineHeight: '16px',
-        desktopSize: 14,
-        desktopLineHeight: '20px',
-    },
-    text2: {
-        mobileSize: 14,
-        mobileLineHeight: '20px',
-        desktopSize: 16,
-        desktopLineHeight: '24px',
-    },
-    text3: {
-        mobileSize: 16,
-        mobileLineHeight: '24px',
-        desktopSize: 18,
-        desktopLineHeight: '24px',
-    },
-    text4: {
-        mobileSize: 18,
-        mobileLineHeight: '24px',
-        desktopSize: 20,
-        desktopLineHeight: '28px',
-    },
-    text5: {
-        mobileSize: 20,
-        mobileLineHeight: '24px',
-        desktopSize: 28,
-        desktopLineHeight: '32px',
-    },
-    text6: {
-        mobileSize: 24,
-        mobileLineHeight: '32px',
-        desktopSize: 32,
-        desktopLineHeight: '40px',
-    },
-    text7: {
-        mobileSize: 28,
-        mobileLineHeight: '32px',
-        desktopSize: 40,
-        desktopLineHeight: '48px',
-    },
-    text8: {
-        mobileSize: 32,
-        mobileLineHeight: '40px',
-        desktopSize: 48,
-        desktopLineHeight: '56px',
-    },
-    text9: {
-        mobileSize: 40,
-        mobileLineHeight: '48px',
-        desktopSize: 56,
-        desktopLineHeight: '64px',
-    },
-    text10: {
-        mobileSize: 48,
-        mobileLineHeight: '56px',
-        desktopSize: 64,
-        desktopLineHeight: '72px',
-    },
 };
 
 export const Text10 = ({dataAttributes, forceMobileSizes, ...props}: TextPresetProps): JSX.Element => {
